+++ packages/platforms/react-native/lib/platform-extensions.tsx
@@ -3,7 +3,6 @@ import React from 'react'
 import { Platform } from 'react-native'
 import NativeBugsnagPerformance from './native'
 import type { ReactNativeAttachConfiguration, ReactNativeConfiguration } from './config'
-import { createAppStartSpan } from './create-app-start-span'
 import type { ReactNativeSpanFactory } from './span-factory'
 
 type NavigationSpanOptions = Omit<SpanOptions, 'isFirstClass'>
@@ -15,13 +14,11 @@ export const platformExtensions = (appStartTime: number, clock: Clock, spanFacto
     return spanFactory.toPublicApi(span)
   },
   withInstrumentedAppStarts: function (App: React.FC) {
-    const appStartSpan = createAppStartSpan(spanFactory, appStartTime)
+    spanFactory.startAppStartSpan(appStartTime)
 
     return () => {
       React.useEffect(() => {
-        if (appStartSpan.isValid()) {
-          spanFactory.endSpan(appStartSpan, clock.now())
-        }
+        spanFactory.endAppStartSpan(clock.now())
       }, [])
 
       return <App />