+++ test/react-native/features/fixtures/scenario-launcher/android/src/main/java/com/bugsnag/reactnative/scenariolauncher/ScenarioLauncherImpl.java
@@ -140,6 +140,9 @@ class ScenarioLauncherImpl {
       editor.putBoolean("useWrapperComponentProvider", configuration.getBoolean("useWrapperComponentProvider"));
     }
 
+    if (configuration.hasKey("scenario")) {
+      editor.putString("scenario", configuration.getString("scenario"));
+    }
 
     editor.commit();
   }
@@ -159,6 +162,7 @@ class ScenarioLauncherImpl {
       startupConfig.putBoolean("autoInstrumentNetworkRequests", sharedPreferences.getBoolean("autoInstrumentNetworkRequests", false));
       startupConfig.putInt("maximumBatchSize", sharedPreferences.getInt("maximumBatchSize", 100));
       startupConfig.putBoolean("useWrapperComponentProvider", sharedPreferences.getBoolean("useWrapperComponentProvider", false));
+      startupConfig.putString("scenario", sharedPreferences.getString("scenario", ""));
       return startupConfig;
     } finally {
       // make sure we don't leave this config around for the next startup
@@ -170,6 +174,7 @@ class ScenarioLauncherImpl {
         .remove("autoInstrumentNetworkRequests")
         .remove("maximumBatchSize")
         .remove("useWrapperComponentProvider")
+        .remove("scenario")
         .commit();
     }
   }